Moved deploy of Sonic SwapX AMO Impl to deployAction
